FaceAttributes flags

+++ Engine/Graphics/Vis.cpp
@@ -27,7 +27,7 @@ Vis_SelectionFilter vis_sprite_filter_2 = {
 Vis_SelectionFilter vis_face_filter = {
     VisObjectType_Face, OBJECT_Any, -1, 0, None};  // 00F93E44
 Vis_SelectionFilter vis_door_filter = {
-    VisObjectType_Face, OBJECT_BLVDoor, -1, FACE_HAS_EVENT, None };  // 00F93E58
+    VisObjectType_Face, OBJECT_BLVDoor, -1, static_cast<int>(FACE_HAS_EVENT), None };  // 00F93E58
 Vis_SelectionFilter vis_sprite_filter_3 = {
     VisObjectType_Sprite, OBJECT_Decoration, -1, 0, ExclusionIfNoEvent};  // 00F93E6C
 Vis_SelectionFilter vis_sprite_filter_4 = {
@@ -349,7 +349,7 @@ void Vis::PickIndoorFaces_Mouse(float fDepth, RenderVertexSoft *pRay,
  * @param model                         Pointer to model to check against.
  * @param reachable_depth               A depth distance for checking interaction against.
  * @param reachable[out]                Whether the model is within the reachable depth specified.
- * 
+ *
  * @return                              Whether the bounding radius of the model is visible within the camera FOV cone.
  */
 bool IsBModelVisible(BSPModel *model, int reachable_depth, bool *reachable) {
